separate mock testing objects and colocate APIs
