+++ blockchain/message_pool/src/lib.rs
@@ -10,6 +10,6 @@ mod msgpool;
 pub use self::block_prob::*;
 pub use self::config::*;
 pub use self::errors::*;
-pub use self::msgpool::*;
 pub use self::msgpool::msg_pool::MessagePool;
 pub use self::msgpool::provider::MpoolRpcProvider;
+pub use self::msgpool::*;
